add pvp result layout
