+++ src/main/java/cn/wthee/pcrtool/adapters/PvpCharacterPageAdapter.kt
@@ -4,7 +4,7 @@ import android.util.SparseArray
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import cn.wthee.pcrtool.ui.tool.ToolPvpCharacterIconFragment
+import cn.wthee.pcrtool.ui.tool.pvp.ToolPvpCharacterIconFragment
 
 class PvpCharacterPageAdapter(activity: FragmentActivity) : FragmentStateAdapter(activity) {
