+++ src/main/java/cn/wthee/pcrtool/data/model/entity/PvpCharacterData.kt
@@ -17,3 +17,11 @@ fun getDefault() = arrayListOf(
     PvpCharacterData(0, 999),
     PvpCharacterData(0, 999),
 )
+
+fun ArrayList<PvpCharacterData>.getIds(): String {
+    var ids = ""
+    for (character in this){
+        ids += character.unitId.toString() + ","
+    }
+    return ids
+}
