+++ src/main/res/layout/fragment_tool_pvp.xml
@@ -26,6 +26,7 @@
         tools:listitem="@layout/item_character_icon" />
 
     <com.google.android.material.button.MaterialButton
+        android:id="@+id/pvp_search"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         style="@style/Widget.MaterialComponents.Button.OutlinedButton"
