+++ src/main/res/layout/fragment_tool_pvp_character_select.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.tool.ToolPvpCharacterIconFragment">
+    tools:context=".ui.tool.pvp.ToolPvpCharacterIconFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/icons"
